requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.4.4
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Werkzeug==2.3.0
